requirements.txt
streamlit
transformers
torch
python-dotenv
sentence-transformers
sentencepiece 
google.generativeai
PyMuPDF
fpdf
nltk
